Redirected all comms output from forked process via file instead of StdOut
This is actually somewhat closer to the original surefire design (version 2.6.X).

Please note that all encoded output from forked process now goes via file system channel
instead of stdout. This really means that the stdout/stderr encoding we're using currently
can really be removed. It may also be possible to simplify the encoding format used in the
file since we avoid any "console" related issues, I have not really checked.
A future version of ForkStarter could consider *any* stdout/stderr output to be client output
or just keep sending the "well formed" client output over the command channel and let
any stdout/stderr traffic be "core dump" related messages that the VM produces from
its deepest bowels before committing harakiri.

But there is, as always a complexity: Shadefire is currently an older version of the
forked process that does not understand the "new" command channel argument.
(And yes, it may be possible to send the command-channel to the old version which
will try to load it as a properties file. This should silently be ignored since
the file is empty).
So currently the forkstarter accepts commands over both channels to remain compatible
with older shadefire. Once ${shadedVersion} understands the new command channel syntax,
we no longer need to accept commands over the stdout channel.

When forking with an updated ForkedBooter, there will be no encoded traffic over
stdout, which should fix problems when VM crashes etc.

Additionally, the file is deleted in the listening process and is created
before the security manager is installed, which should reating compatibility
with security managers.